Change float literal to use double-precision
Single-precision floats are a bit restrictive, fairly simple
move from using `Float` to `Double`.

+++ src/Angle/Lex/Token.hs
@@ -174,7 +174,7 @@ tokDigits = some tokDenaryDigit
 
 
 -- TODO: Might be able to use a standard read?
-tokFloat :: Scanner Float
+tokFloat :: Scanner Double
 tokFloat = do
   negve <- optional (char '-')
   f <- tokDigits
